✨ Build approximate VM spec when reverting to imported snapshots
This commit introduces the ability to build an approximate VM spec when
reverting to imported snapshots. The spec is constructed based on the
snapshot's metadata and some fields from the state of the VM in VC,
allowing for a more accurate representation of the VM's state at the
time of the snapshot.

This is done so that the VM spec can reach to a state that likely defines
how the VM was configured when the Snapshot was taken.

Signed-off-by: Nabarun Pal <nabarun.pal@broadcom.com>