Fix use of old condition constant
This patch fixes an issue where the old OVFReady condition was used
in the VM group placement test instead of the new HardwareReady
condition. This only occurred because the PR that made the constant
change was merged without any conflicts, even though it probably should
have failed due to the PR merged ahead of it that used this condition
constant that had been changed.