Disable fast-deploy for instance storage VMs
This is just a quick temp fix to disable fast-deploy when a VM
has instance storage. This depends on a implementation detail in
that the "first" reconcile will add the class's instance storage
volume to the VM Spec so later reconciles can detect that and
once the instance storage PVCs are bound, create the VM without
fast-deploy.

This will be fixed for real once we can adjust placement to
support a specific, already selected host for later placements
where we just need the datastores.