Support non-friendly name for FastDeploy+vSAN
This patch updates the way VMs are provisioned on vSAN using
Fast Deploy. Instead of using the VM folder's friendly name, it
is first converted to the UUID path. Otherwise there is a bug in
vpxd that ignores the friendly path during VM create and decides
the folder does not exist, creating yet another new folder.