Disable FastDeploy+Direct / NVRAM file support
This patch temporarily disables FastDeploy+Direct mode as well as
support for FastDeploy with NVRAM files. They may both be re-enabled
by setting a single flag to true.

This patch is required to mitigate a bug in vpxd on main related to
vSAN datastores. Currently they do not support creating VMs using an
explicit path.