Delete dir created during VM deploy
This patch updates the Fast Deploy workflow to track the path of
the directory created during VM provisioning and to delete that
directory when a VM is being deleted. If the directory no longer
exists, then no error will occur.

This patch is required to mitigate a bug in vpxd on main related to
vSAN datastores. Currently they do not support creating VMs using an
explicit path.