Log VM recent tasks / ignore certain tasks
This patch adds logging to record the VM's recent tasks when a
VM is reconciled. The log level depends on the task's state. If
running or in error, the standard level, 2, is used. Otherwise
the task is logged at level 4.

This patch also updates HasVMRunningTask to ignore certain,
well-known tasks that should not prevent VM Op from reconciling
desired state.