Disable ExternalID for NetOP interfaces
The fix to enable setting the EthCard ExternalID on non-NSXT DVPDs
(NetOP in our case) is partially dependent on ESX host changes, and
for upgrades the host version may very much lag of VC/VMOP versions.

The fallout from this is that like before we cannot easily uniquely
identify NetOP interfaces so we'll fallback to the old best-effort
matching by backing.