:seedling: Prevent snapshot and revert of VKS nodes (#1097)
This change modifies the VM snapshot webhook to reject snapshot
requests on VKS nodes.  This is done by checking for CAPI added labels
on the VKS nodes which is already a contract / pattern that is
followed in VM operator for other workflows (e.g., backup).

Since webhook validation based on labels is always race-y, this change
also modifies the snapshot reconciliation workflow in the provider to
skip taking snapshot of VKS nodes.

Same restrictions are also enforced on revert workflows.