Add MacAddress to VM InterfaceSpec
For VPC SubnetPort set the InterfaceSpec IP and MAC address
on the SubnetPort Spec.

Fix not honoring the inteface spec Gateway4/6 unless Addresses
were also specified.

Fix Gateway4/6 documentation incorrectly saying these fields are
to include the network prefix.

While here, write the code for VDS but disallow it in the webhook
until we decide if we really want to support it there.