:seedling: Enqueue reconcile on snapshot and root snapshot updates
This change modifies the watcher to enqueue an event when the
snapshot (the current snapshot) and the root snapshot (list of root
snapshots) of a VM change.  This allows us to not worry about updating
/ patching the snapshot tree once we are done with the snapshot or
revert operation.  Once the snapshot is done, it will queue a
reconcile and refresh the snapshot tree.

Testing Done:
- Unit tests
- Verified on a testbed that the watcher gets called when a snapshot
is created on vCenter directly.  The list of properties returned by
vCenter also include the (current) snapshot, and root snapshots.