:seedling: Remove the zone annotation from VM backup
Today, a VM being backed up includes the zone annotation.  As a
result, if a VM is being restored from backup in an environment where
the zone doesn't exist, the VM registration will fail.

This problem is much more severe for the failover scenarios since the
primary and secondary sites are different and can have distinct zones.

Testing Done:
- Modified the existing unit tests to validate that the label is unset
- Validated in a real env that the label is dropped from the backup