:sparkles: Support VM-VM anti-affinity at Zonal level (#1119)
This change introduces the support for specifying anti-affinity among
VMs with respect to zones.  This allows a VM to specify an effective
policy that can say "do not place this VM in the same zone with other
VMs that have these labels".

One change to call out is that now we are sending in a list of tags
that should be attached to the VMs even if a VM doesn't specify any
affinity policy at all.  This is so that when we stat supporting
dynamic policy evaluations, the tags specified by day-2 policies are
already present on the VMs.

Testing Done:
- Other than the unit tests here, I have evaluated this in a real
environment where the anti-affinity policy is being respected.

Signed-off-by: Arunesh Pandey <aruneshpa@gmail.com>