🌱 Emit a condition for a failed snapshot revert operation
When a snapshot revert operation fails, we now emit a condition on the
VirtualMachine object to indicate the failure. This condition provides
details about the failure, allowing users to understand the reason behind
the unsuccessful revert attempt.

Once the revert operation fails, users will see a condition named
`VirtualMachineSnapshotRevertSucceeded` set to `False` on the VirtualMachine
object. The condition's message will contain information about the error
encountered during the revert process.

Also, enhances documentations for the `VirtualMachineSnapshotRevertInProgressAnnotationKey`
to clarify its behavior in failure scenarios.

Signed-off-by: Nabarun Pal <nabarun.pal@broadcom.com>