🌱 Reverts a VM to the same power state as that of the snapshot
For non-imported snapshots, so far the power state was being taken
from the backed up VM YAML in ExtraConfig for the snapshot. This
might not be the correct power state because, for example, in the
case when one takes a snapshot of a Powered On VM without it's memory
state, the snapshot power state would be off. The VM's resultant
power state post revert should also be off.

Signed-off-by: Nabarun Pal <nabarun.pal@broadcom.com>