:seedling: Apply VM name label on the snapshot custom resource (#1137)
This change introduces a new label on a `VirtualMachineSnapshot`
resource that points to the VM custom resource for which the snapshot
is taken.  This label is added in the mutation webhook when the
resource is created.  The value of the label is extracted from
snapshot resource's `spec.vmRef.name`.

This allows clients (such as VCFA) to query the snapshots resources
corresponding to a given virtual machine.

The mutation webhook _only_ intercepts the Create operations since we
do not support changing the `vmRef` of an existing snapshot resource.

To enforce this label never changing, introduce a validation in the
validation webhook.