Convert VM in extra config of snapshot to newest APIVersion (#1130)
There is chance that the VM Spec's content of a Snapshot doesn't have the VM with latest apiVersion. When revert to snapshot, this PR extracts and converts the spec to latest (v1alpha5) version of VM.
Also fix the powerstate, it should match what the `vmsnapshot.spec.memory` is set, not the state of the vm.


Signed-off-by: Lubron Zhan <lubronzhan@gmail.com>