✨ Add support for Unmanaged snapshots in VM snapshot references
This makes references to VM snapshots use a new struct
`VirtualMachineSnapshotReference` which can refer to both Managed
and Unmanaged snapshots. The existing fields `CurrentSnapshot` and
`RootSnapshots` in `VirtualMachineStatus` and `Children` in
`VirtualMachineSnapshotStatus` have been updated to use this new struct.

This also adds conversion logic to convert between v1alpha4 and v1alpha5
APIs. The conversion logic assumes that all snapshot references in
v1alpha4 are to Unmanaged snapshots. The conversion logic has been
tested in the integration tests.

Note: VM's spec.CurrentSnapshot is unchanged as it is not a supported
mechanism to revert to anything other than Managed snapshots.

This also updates the `SyncVMSnapshotTreeStatus` function to
appropriately set the new fields, even in case of Unmanaged snapshots.

Signed-off-by: Nabarun Pal <nabarun.pal@broadcom.com>