:seedling: Allow privileged users to unset the zone annotation on the VM (#1156)
Backup Restore and Disaster recovery workflows rely on the backup
YAML packaged in the VM.  The backup YAML does not contain the zone
annotation because the VM can be restored, or failed-over on a
different zone than where it was created.

During restore, when the VM is applied with empty zone label from the
backup, VM operator should allow that request.  During reconciliation,
if the VM exists, we perform a reverse lookup of the zone from the
resource pool where the VM exists.  VM operator then also adds the
zone label on the VM.