Return no requeue error when group member is NotFound or NotMember (#1157)
This PR updates the VMG controller to return a non-requeue error when the member object is not found or not a group member (i.e. member.Spec.GroupName != VMG.Name). This ensures the `lastUpdatedPowerState` annotation is not set in the group's status until all its members are found and linked to the group.

While here, removing the `logger.Error()` as the returned error is already logged by the controller-manager.