🐛 Clear out the snapshot revert condition on no op
When a user tries to revert to a snapshot, a condition
is set on the VM to indicate the status of the revert
operation.

If for some reason, the revert operation fails to
complete, the condition remains on the VM. The user
can either take mitigation steps or abort the revert
operation. When a user aborts the revert operation,
we should clear out the condition.

Signed-off-by: Nabarun Pal <nabarun.pal@broadcom.com>