:seedling: Support explicit zone specification from group members (#1171)
Placement of a VM is driven via the VirtualMachineGroup if the VM is
part of a group (specified by spec.groupName).  This means we don't
send in the list of placement candidate resource pools to DRS when
asking for placement recommendations.  DRS returns recommendations
based on the hardware and the placement policies specified with the
VM.

However, if a group member wants to explicitly be placed in a
particular zone, they should be able to specify the zone label on the
member.  This change adds support for that.  This will be used by VKS
to support scale-outs of node pools.