:seedling: Remove all unused fields and types from v1alpha3 and v1alpha4 (#1158)
We shipped some types and fields in older versions of the
API (v1alpha3 and v1alpha4) without the backing implementation.  These
were never meant to be released in these API versions.

This change removes these types and fields.  Technically, this is a
breaking change, but since no one was using these types (because they
were never implemented) on these versions, it is safe to remove.

Note that this change does not solve the issue of CRDs being installed
in v1alpha5 version. That is a known issue and is being worked on
separately.