Allow VM Spec.Bootstrap to be mutable
Previously we would restrict the bootstrap to only be changed with the
VM was "off", but there isn't really a good reason to do so. The
customization will be applied on the next powering on transition.
Instead, just restrict the bootstrap provider type from being changed.

Remove a bogus "network can't change while powered on" test: this
was just testing network mutability or not since we had no network
related checks for just powered on.