Remove GOSC VCD Parity fields from VM when not enabled
Remove the fields added in 7b12de5 when capability is not enabled.
Update things to allow for removing of nested fields under spec
and status. Use dot field separators in the tests since that is
a little easier to grab as-is from like k explain, and to reduce
the overall chances of fat fingering anything.