Update VMI DiskName logic
This patch updates how the name of a VMI disk is calculated. It now uses
the ElementName from the OVF instead of the disk ID. This is the same value
used as the Description.Label field when an OVF is converted to a
ConfigSpec, allowing a mapping between the two.

This patch also fixes an issue with how the disks were discovered from an
OVF. Previously they were discovered by iterating over the OVF's DiskSection.
However, this is just a table-of-contents. It does not actually represent
what disks the VM in the OVF may be using. There could be a disk for each
OS architecture, and the VM may only be using one of them. This patch correctly
parses disks based on the ones used by the VM and then looks up the disk info
from the disk section.

Finally, this patch added the UUID field to the UnmanaagedVolumeClaim type
in the VM Op APIs. This is to allow mutating in a new UUID for disks added
from VMIs to a VM. Since disks may be created on a VM with a well-known UUID,
this change allows VM Operator to maintain a mapping between the disk in the
VMI all the way through to the disk in the VM, first by the label on the disk
from OVF-->ConfigSpec and then by the disk UUID once the VM is created.