:bug: Use InstanceUUID for batch attachments instead of Bios UUID
We made a decision to pivot to vCenter instance UUIDs in the batch
attachment spec in order to uniquely identify a VM in a vCenter
inventory.  But we were still specifying the Bios UUID in the batch
attachment spec when creating it.  This resulted in the VM lookup
failing in CSI since they have not implemented the fallback to lookups
by Bios UUID.

This change modifies the controller to set the NodeUUID with the
instance uuid of the VM.  Also change associated tests.