Restrict watched guest properties
This patch updates the async watcher to no longer watch the prop
path "guest.ipStack". This caused too much churn due to VKS VMs
constantly adding/removing IP routes due to pods.

It would be ideal if we could watch just the first/second IP route,
but it is not possible to watch indexed property paths,