Use priority queue / ctrl-runtime v0.22.0 / 30m resync
This patch updates the VM controller to use the controller-runtime
priority queue feature by default in combination with v0.22.0 of
controller-runtime. This feature ensures that reconcile requests that
are not part of an initial list or resync have lower priority than those
requests enqueued for any other reason. This should allow VM Operator to
handle VM reconciliation when the queues are do large due to a high number
of VMs.

This patch also changes from a 10m resync to 30m, relying
on the async watcher to enqueue VMs for reconciliation in a timely
manner. This also means other controllers need to ensure they are properly
using requeues / mappers to ensure objects are not relying on resync to
refresh status in a timely manner.