Only set at most one IP address in the SubnetPort Spec
For addresses that set in our interface spec, only set the first
valid one in the SubnetPort since despite being list, that is all
VPC currently supports.

Handle the case where IP address our set in our interface spec
but the gateway is not. In that case, use the gateway that was
provided by the network provider.

I modeled our IPConfig results after what the network interface CR
Status have, but that is getting cumbersome for our needs here. It
is kind of confusing to see multiple gateways set. We should hoist
out the gateways and just separate out the IPv4 and IPv6 addresses.