Add deviceKey for hardware status controllers (#1241)
This patch adds the deviceKey to VM status's controllers. The deviceKey
is used as controllerKey to generate CNSNodeVMBatchAttachment volumes.
Given a controller type and a bus number, a controller is still pending
to created and attached to the VM if the VM status hardware controllers
does not contain a controllerStatus with its controllerKey. Therefore,
the vm.spec.volume with the missing controllerKey is skipped
reconciling until the defined controller is attached and the vm status
is updated.