Use VMSetRP Status to determine if VM cluster module exists
During VM creation, we would check if all the VMSetRP's cluster modules
exists on VC. It doesn't really make sense to check that all of the
cluster modules exist since a VM can only be assigned at most one. Also,
due to the cluster module API, the DoesModuleExist() helper has to
list all of the modules to find the one it cares about which is not very
efficient.