Do not reconcile VMs if only status updated
This patch updates the VM controller to filter any UPDATE
events unless the VM's generation has changed. This ensures
that all events related purely to status updates do not cause
additional reconciles.

When there is an occasion to need to reconcile a VM due to what
would have been a status change, it is possible to use the VM's
channel-based source to enqueue an event.