Report VM's used storage as the requested capacity
We are currently reporting the VM's total storage as the requested
capacity during the async quota flow when publishing to an inventory
content library. We should instead be reporting the used storage. This
change updates the publish request controller to sum up the values in
status.storage.used and report that value as the requested capacity.