Skip CD-ROM validation for unknown power state VM
The webhook validation was added to deny any changes to the CD-ROM
controller placement spec when the VM is powered On. When a new VM
is created, it goes through the schema upgrade phase first which
triggers an update operation to the VM after it adds the upgrade-to
annotations to the VM. The mutation webhook will try to fill out
any missing placement information; however, it is denied due to
old VM spec and new VM spec both have power state sets to on. This
change is to allow the first time mutation after the schema upgrade.
Therefore, when vm status power status is unknown, we skip the check.