Update Policy Evaluation Controller to add compute policies correctly (#1284)
This PR updates the policy evaluation controller to apply compute polices as follows:
- Automatically apply **mandatory** polices if matched (optional polices are added explicitly) 
- For explicit polices defined in `PolicyEvaluation.Spec.Policies` (populated from `VM.Spec.Polices`), add them only if matched (regardless of mandatory or optional mode)

Also, this PR renames the existing `reconcileMatchingPolicies` function to `reconcileMandatoryPolicies` so it aligns better the purpose and the other `reconcileExplicitPolicies`  function called together by the Reconciler.

While here, adding the missing `patch` and `update` controller RBAC permission for `policyevaluations/status`.