Add VM condition to reflect VMIC state
The presence of the vmicache.vmoperator.vmware.com/name label is the
only apparent indication that the VM create is waiting for the VMIC
in the selection location to be ready. Add a VM condition for that,
like:

  - lastTransitionTime: "2025-11-20T18:15:32Z"
    message: cached files not ready
    reason: NotReady
    status: "False"
    type: VirtualMachineConditionImageCacheReady

And when the cache is ready in the location, from the VMIC:

  locations:
  - conditions:
    - lastTransitionTime: "2025-11-20T18:20:43Z"
      message: ""
      reason: "True"
      status: "True"
      type: Ready
    datacenterID: datacenter-40
    datastoreID: datastore-48

That will be reflected on the VM:

  - lastTransitionTime: "2025-11-20T18:20:44Z"
    message: VirtualMachineImageCache location datacenter-40:datastore-48
             was ready at 2025-11-20T18:20:43Z
    reason: "True"
    status: "True"
    type: VirtualMachineConditionImageCacheReady

This is helpful to later determine VM provisioning delays.

Update the VirtualMachineImageReady condition to not mirror the VMI
Ready condition exactly, but to include its ready timestamp in the
condition message:

  - lastTransitionTime: "2025-11-20T18:15:32Z"
    message: VirtualMachineImage was ready at 2025-11-17T23:42:30Z
    reason: "True"
    status: "True"
    type: VirtualMachineImageReady