Track VMPub clone tasks based on activationID (#1289)
This PR updates the VirtualMachinePublishRequest controller
to more accurately track which task is associated with a VirtualMachine
clone. We are currently attempting to associate a task with the
target template based on the result. But the template may not be
available in inventory, and the result may be nil. We now set the
VMPub UID as part of the vimtypes.ID on the context passed into the
call to clone. This allows us to filter tasks based on the presence
of the UID in the activationID for the clone task.