Cancel disk promotion task during VM delete
A disk promotion task will exclusively lock the VM,
which will block delete. Since the VM is going to be
deleted anyways cancel the promotion task.

Reorder a bit of how some things are checked during delete.
Since we have to grab the VM properties earlier, move the
paused check to before canceling a task.

Fix a logger panic when a Task completedTime is nil