Emit condition when power state synced
This patch updates the VM API with a condition,
VirtualMachinePowerStateSynced, that is only true when the
VM's observed power state matches its desired power state. This is
incredibly useful for a number of reasons, but a clear use case is
for debugging when we want to know when a VM was powered on.