Add detaching suffix to vm.status.volumes (#1320)
For batchAttachment workflow, CSI now adds `:detaching` 
suffix to volume name when a volume is being detached. 
But it only adds it once it finished a detaching request 
to CNS, which could take up to minutes. This means there 
is a time period that the volume is not in vm.spec.volumes 
but still in batchAttachment.status.volumes, we still 
want to capture those volumes' status inside 
vm.status.volumes.