Handle vSphere Policy Evaluations during group placement
Group placement uses much of the code with the non-group
VM create code path, but where policies were done was
not a part of the group uses, since that is done closer
to single VM placement.

Since the policy eval object will need to be reconciled
before it is ready, stop failing at the first error, so
all the needed objects will be created.

Set the VM group member PlacementReady condition when
things fail to give more info. Note that since when
the regular VM reconcile bails for not yet ready grouped
VMs, the VM itself will also have that info but add to the
group for convenience.

Update the PolicyEval to also check for Ready condition