Default in both interface Spec network APIVersion and Kind
I think my thought at the time was to only mutate this in if
both APIVersion and Kind are unset, since if one is set, the
default network may clash. But heterogeneous network env are
not really in the immediate future.

This needs to be followed up by a validation change to only
allow expected APIVersion and Kind.