Patch out VMIC loc w empty profileID
Check if any of the spec.locations or status.locations have an empty
profileID field and remove those entries if they do. This is a
side-effect of a weird order-of-operations on Supervisor where the Fast
Deploy FSS could get enabled prior to the VMIC CRD being updated,
causing the old version to be used before the field was introduced.