:sparkles: Sanitize VMs during Unregister
VM operator persists various metadata to the vCenter VM, predominently in
its ExtraConfig.  This is done so that we even if we restart VM
operator, we have crucial information stored with the VM.  In other
cases, this is done so that external vendors that only deal at the VMX
file level can persist this data.

We also have a workflow where we need to allow for deletion of a VM
custom resource without the underlying vCenter VM being deleted.  This
is typically done when an import operation can't be completed due to
some errors and is rolled back.  In this case, we would want the VM to
be free of any metadata stored in the VM so that we return it in the
same state as we adopted it.

This change implements this behavior.  Note that this change only
cleans up the ExtraConfig and the managedBy properties.  VM operator
also adds tags to VMs.  But that is a little contentious, so I am
leaving that for now.

Testing Done:
- Deployed this change on a testbed
- Created a VM and verified that it has the ExtraConfig
- Added the skip-platform-delete annotation
- Deleted the VM CR
- Verified that the vCenter VM doesn't have the ExtraConfig