Include policy tags in initial VM ConfigSpec
To ensure proper placement, add the policy tags to the initial
ConfigSpec TagSpecs. The VM will also be created with the correct
tags, but this is less of an issue since we would have later
added the tags to the VM.

Fix the handling of the ExtraConfigPolicyTagsKey: this should
contain the "shouldBe" tags, that is the tags from the policy
eval since that is the desired tags.