Skip CD-ROM reconcile when controllers pending (#1338)
Add early return in UpdateCdromDeviceChangesWithSharedDisks() to skip
CD-ROM reconciliation if IDE or SATA controller changes are in the
configSpec. This allows controllers to be created first, then CD-ROMs
are reconciled in a subsequent loop.

This patch also fixes schema upgrade check from oldVM to newVM webhooks
and SetPVCVolumeDefaults() function to override the default
configuration for a volume based on its application type.