Allow VM Op to bypass vol validation
This patch allows VM Op's svc account to bypass volume validation.
There are two design flaws related to how VM Op does backfill for
managed disks as well as
https://github.com/vmware-tanzu/vm-operator/pull/1338.

The current method for backfilling managed disks expects the disks to:

* be legacy disks (i.e. come from CnsNodeVmAttachment) instead of from
  atch attach objects
* be attached at all!

The schema upgrade for disks should perform the same logic as the
mutation webhook. Instead, because the mutation webhook is relied on for
this exclusively, it is subject ot a pretty big design flaw. It has the
following code:

    if (pkgcfg.FromContext(ctx).Features.VMSharedDisks ||
        pkgcfg.FromContext(ctx).Features.AllDisksArePVCs) &&
        vmopv1util.IsVirtualMachineSchemaUpgraded(ctx, *modified) {

This means the mutation webhook will not mutate in unit numbers, etc.
into the VM until its schema has been upgraded.

But the validation webhook, which runs after mutation, has this:

    ctx *pkgctx.WebhookRequestContext,
    vm, oldVM *vmopv1.VirtualMachine) field.ErrorList {

    if oldVM != nil && !vmopv1util.IsVirtualMachineSchemaUpgraded(ctx, *oldVM) {
        return nil
    }

It removed the priv check that allows VM Op to skip validation, and thus
will run AFTER the VM has been schema upgraded.

This means that by design, the above PR will fail every time.

How did the original PR pass the sniff test? Because I myself left a
note near a change the PR's author made:

    if ctx.IsVMOperatorAccount {
        // TODO(akutz): This can be removed after we merge the implementations
        // of reconcileBackfillUnmanagedDisks and reconcileSchemaUpgrade
        // in vmprovider_vm.go into a single patch of the VM CRD.

The privilege check that allowed the VM Op svc account to bypass volume
validation was removed because we did in fact combine the two backfill
implementations like the note said.

However, I wrote that TODO before I realized we did not actually
backfill the managed volumes as part of schema upgrade. Since that only
happens as part of mutation, we are back to the fatal flaw in the
aforementioned PR.

This patch reintroduces the change that allows the VM Op svc acct to
skip volume validation. The long-term fix is to due to the managed
volume backfill in the schema upgrade location where it belongs.