Support feature downgrade
This patch cleanly supports IsVirtualMachineSchemaUpgraded when a
capability is later disabled. It achieves this by ensuring that even
if the activated feature version is no longer equal to the feature
version on an object, the activated version's bits are checked against
the object's version. If the object's version still contains all of the
activated version's bits, then it's known the object was upgraded to
the activated version at some point, even if the latter changed.