Do not set PlacementReady condition if async VM create fails
If the create fails, this will cause the PlacementReady condition to
flip flop: before kicking off async create, as a part gathering the
args needed to create the VM, the PlacementReady condition will be
set to true, which will update the condition when that reconcile
returns, and then the async create failure will update it.

The Created condition is already updated with the failure so the
cause is still reported.